+++ apps/creator-app/src/ducks/designer/reference/reference.reducer.ts
@@ -1,3 +1,4 @@
+import { Utils } from '@voiceflow/common';
 import { Actions } from '@voiceflow/sdk-logux-designer';
 import * as Normal from 'normal-store';
 import { reducerWithInitialState } from 'typescript-fsa-reducers';
@@ -24,5 +25,29 @@ export const referenceReducer = reducerWithInitialState<ReferenceState>(INITIAL_
       refererIDsByResourceIDMap: { ...state.refererIDsByResourceIDMap, ...cache.refererIDsByResourceIDMap },
       referenceIDsByResourceIDMap: { ...state.referenceIDsByResourceIDMap, ...cache.referenceIDsByResourceIDMap },
       referenceIDsByReferrerIDMap: { ...state.referenceIDsByReferrerIDMap, ...cache.referenceIDsByReferrerIDMap },
+    };
+  })
+  .case(Actions.Reference.DeleteMany, (state, { data }) => {
+    const resourceIDs = data.referenceResources.map((resource) => resource.id);
+    const referenceIDs = Utils.array.unique([
+      ...data.references.map((reference) => reference.id),
+      ...resourceIDs.flatMap((resourceID) => state.referenceIDsByResourceIDMap[resourceID] ?? []),
+      ...resourceIDs.flatMap((resourceID) => state.referenceIDsByReferrerIDMap[resourceID] ?? []),
+    ]);
+
+    return {
+      resources: Normal.removeMany(state.resources, resourceIDs),
+      references: Normal.removeMany(state.references, referenceIDs),
+      blockNodeResourceIDs: Utils.array.withoutValues(state.blockNodeResourceIDs, resourceIDs),
+      triggerNodeResourceIDs: Utils.array.withoutValues(state.triggerNodeResourceIDs, resourceIDs),
+      resourceIDsByDiagramIDMap: data.referenceResources.reduce((acc, resource) => {
+        if (!resource.diagramID) return acc;
+
+        return { ...acc, [resource.diagramID]: Utils.array.withoutValue(acc[resource.diagramID] ?? [], resource.id) };
+      }, state.resourceIDsByDiagramIDMap),
+      resourceIDsByRefererIDMap: Utils.object.omit(state.resourceIDsByRefererIDMap, resourceIDs),
+      refererIDsByResourceIDMap: Utils.object.omit(state.refererIDsByResourceIDMap, resourceIDs),
+      referenceIDsByResourceIDMap: Utils.object.omit(state.referenceIDsByResourceIDMap, resourceIDs),
+      referenceIDsByReferrerIDMap: Utils.object.omit(state.referenceIDsByReferrerIDMap, resourceIDs),
     };
   });