+++ apps/realtime/src/reference/reference.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ReferenceORM, ReferenceResourceORM } from '@voiceflow/orm-designer';
+import { IntentORM, ProjectORM, ReferenceORM, ReferenceResourceORM } from '@voiceflow/orm-designer';
 
 import { ReferenceLoguxController } from './reference.logux.controller';
 import { ReferenceService } from './reference.service';
@@ -7,7 +7,7 @@ import { ReferenceCacheService } from './reference-cache.service';
 
 @Module({
   exports: [ReferenceService],
-  providers: [ReferenceORM, ReferenceResourceORM, ReferenceService, ReferenceCacheService],
+  providers: [ReferenceORM, IntentORM, ProjectORM, ReferenceResourceORM, ReferenceService, ReferenceCacheService],
   controllers: [ReferenceLoguxController],
 })
 export class ReferenceModule {}