+++ libs/realtime-sdk/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.417.1](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.417.0...@voiceflow/realtime-sdk@1.417.1) (2024-06-17)
+
+**Note:** Version bump only for package @voiceflow/realtime-sdk
+
 # [1.417.0](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime-sdk@1.416.0...@voiceflow/realtime-sdk@1.417.0) (2024-06-14)
 
 ### Features